+++ src/scripts/lobby/components/AvatarSelect.js
